fix : 팀원평균점수 평가가 덜 된 경우 null

+++ src/main/java/com/devee/devhive/domain/project/member/controller/ProjectMemberController.java
@@ -37,7 +37,7 @@ public class ProjectMemberController {
 
     for (Project project : projects) {
       if (project.getStatus() == ProjectStatus.COMPLETE) {
-        long memberCount = projectMemberService.getProjectMemberByProjectId(project.getId()).size();
+        int memberCount = projectMemberService.getProjectMemberByProjectId(project.getId()).size();
         ProjectHistoryDto projectHistory = ProjectHistoryDto.of(project.getName(),
             projectReviewService.getAverageTotalScoreByTargetUserAndProject(userId, project.getId(), memberCount));
         projectHistoryList.add(projectHistory);
